% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_banded_cov.R
\name{fct_banded_cov}
\alias{fct_banded_cov}
\title{Create Banded Covariance Matrix}
\usage{
fct_banded_cov(dim, type = c("large", "small"), perc_large = 0.01)
}
\arguments{
\item{dim}{integer dimension of matrix}

\item{type}{large or small}

\item{perc_large}{if large, percentage of large entries}
}
\value{
matrix
}
\description{
Create Banded Covariance Matrix
}
