% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_boltzmann_pert.R
\name{fct_boltzmann_pert}
\alias{fct_boltzmann_pert}
\title{Boltzmann Perturbation}
\usage{
fct_boltzmann_pert(lik)
}
\arguments{
\item{lik}{vector of likelihoods}
}
\value{
perturbation fraction
}
\description{
Boltzmann Perturbation
}
