% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_conv_check.R
\name{fct_conv_check}
\alias{fct_conv_check}
\title{fct_conv_check}
\usage{
fct_conv_check(
  x,
  y,
  k,
  N,
  clust_assign,
  selection,
  alpha,
  beta,
  y_sparse,
  rank,
  max_rank,
  ll
)
}
\description{
fct_conv_check
}
