% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_split.R
\name{fct_data_split}
\alias{fct_data_split}
\title{Split data for Cross Validation}
\usage{
fct_data_split(x, y, val_frac)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{val_frac}{fraction for validation}
}
\value{
training and testing sets
}
\description{
Split data for Cross Validation
}
