% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_full_model_error.R
\name{fct_full_model_error}
\alias{fct_full_model_error}
\title{Compute Model Error}
\usage{
fct_full_model_error(x, y, k, model)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{number of clusters}

\item{model}{model object from hthmix}
}
\value{
error
}
\description{
Compute Model Error
}
