% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gamma.R
\name{fct_gamma}
\alias{fct_gamma}
\title{Compute Posterior}
\usage{
fct_gamma(x, y, k, N, clust_assign)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{clusters}

\item{N}{doc}

\item{clust_assign}{doc}
}
\value{
doc
}
\description{
Compute Posterior
}
