% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_global_opt.R
\name{fct_global_opt}
\alias{fct_global_opt}
\title{Global Likelihood Update}
\usage{
fct_global_opt(ll_i, global_opt_ll)
}
\arguments{
\item{ll_i}{current likelihood}

\item{global_opt_ll}{global likelihood}
}
\value{
updated global likelihood
}
\description{
Global Likelihood Update
}
