% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_initialize.R
\name{fct_initialize}
\alias{fct_initialize}
\title{Initial Assignment}
\usage{
fct_initialize(k, N)
}
\arguments{
\item{k}{number of clusters}

\item{N}{number of observations}
}
\value{
assignment vector
}
\description{
Initial Assignment
}
