% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_lam_coef.R
\name{fct_lam_coef}
\alias{fct_lam_coef}
\title{Compute penalization coefficient}
\usage{
fct_lam_coef(x, sigma, m, p)
}
\arguments{
\item{x}{matrix}

\item{sigma}{value}

\item{m}{integer}

\item{p}{integer}
}
\value{
penalization coefficient
}
\description{
Compute penalization coefficient
}
\examples{
fct_lam_coef(matrix(1:9, nrow = 3), sigma = 1, m = 3, p = 3)
}
