% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_log_lik.R
\name{fct_log_lik}
\alias{fct_log_lik}
\title{Compute Log Likelihood}
\usage{
fct_log_lik(mu_mat, sig_vec, y, N, m)
}
\arguments{
\item{mu_mat}{matrix}

\item{sig_vec}{vector}

\item{y}{value}

\item{N}{observations}

\item{m}{dimension}
}
\value{
array
}
\description{
Compute Log Likelihood
}
