% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pert_assign.R
\name{fct_pert_assign}
\alias{fct_pert_assign}
\title{Perturb Assignments}
\usage{
fct_pert_assign(clust_assign, lik_store, N, k)
}
\arguments{
\item{clust_assign}{current assignment}

\item{lik_store}{likelihood vector}

\item{N}{observations}

\item{k}{clusters}
}
\value{
updated assignment
}
\description{
Perturb Assignments
}
