% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pi_vec.R
\name{fct_pi_vec}
\alias{fct_pi_vec}
\title{Compute pi vector}
\usage{
fct_pi_vec(clust_assign, k, N)
}
\arguments{
\item{clust_assign}{current cluster assignment}

\item{k}{clusters}

\item{N}{observations}
}
\value{
pi vector
}
\description{
Compute pi vector
}
