% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rank_var.R
\name{fct_rank_var}
\alias{fct_rank_var}
\title{Title}
\usage{
fct_rank_var(X_k, Y_k, n_k, p, m)
}
\arguments{
\item{X_k}{matrix}

\item{Y_k}{matrix}

\item{n_k}{observations}

\item{p}{dimension}

\item{m}{dimension}
}
\value{
estimated noise and rank
}
\description{
Title
}
