% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sarrs.R
\name{fct_sarrs}
\alias{fct_sarrs}
\title{Compute SARRS algorithm}
\usage{
fct_sarrs(
  Y,
  X,
  r,
  lam,
  alpha = 2 * sqrt(3),
  beta = 1,
  sigma,
  ptype = "grLasso",
  y_sparse = TRUE
)
}
\arguments{
\item{Y}{response matrix}

\item{X}{design matrix}

\item{r}{rank}

\item{lam}{penalization param}

\item{ptype}{penalty type}

\item{V0}{v0}
}
\value{
estimate of coefficients
}
\description{
Compute SARRS algorithm
}
