% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sim_mixrrr.R
\name{fct_sim_mixrrr}
\alias{fct_sim_mixrrr}
\title{Simulate SARRS setting}
\usage{
fct_sim_mixrrr(n, a_rows, p, m, r, rho, sigma, b)
}
\arguments{
\item{n}{sample size}

\item{a_rows}{non-zero coefficient rows}

\item{p}{number of x features}

\item{m}{number of y features}

\item{r}{rank}

\item{rho}{correlation coefficient}

\item{sigma}{error sd}

\item{b}{coefficient}
}
\description{
Simulate SARRS setting
}
