% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sim_tsvd.R
\name{fct_sim_tsvd}
\alias{fct_sim_tsvd}
\title{Simulate SARRS setting}
\usage{
fct_sim_tsvd(
  n = 100,
  p = 200,
  m = 250,
  b = 1,
  rank = 1,
  h = 0.2,
  case = "independent"
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of x features}

\item{m}{number of y features}

\item{b}{coefficient}

\item{a_rows}{non-zero coefficient rows}

\item{r}{rank}

\item{rho}{correlation coefficient}

\item{sigma}{error sd}
}
\description{
Simulate SARRS setting
}
