% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate.R
\name{fct_simulate}
\alias{fct_simulate}
\title{Title}
\usage{
fct_simulate(
  N = 200,
  k = 3,
  rho = 0,
  sigma = 1,
  p = 100,
  m = 50,
  s_x = 10,
  s_y = 15,
  rank = 2,
  b = 1,
  type = "2",
  h = 0.2,
  case = "independent"
)
}
\arguments{
\item{N}{doc}

\item{k}{doc}

\item{rho}{doc}

\item{sigma}{doc}

\item{p}{doc}

\item{m}{doc}

\item{s_x}{doc}

\item{s_y}{doc}

\item{rank}{doc}

\item{b}{doc}
}
\value{
doc
}
\description{
Title
}
