% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update_clust.R
\name{fct_update_clust}
\alias{fct_update_clust}
\title{Update Cluster}
\usage{
fct_update_clust(N, gamma)
}
\arguments{
\item{N}{observations}

\item{gamma}{posterior}
}
\value{
updated cluster assignments
}
\description{
Update Cluster
}
