% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_weighted_ll.R
\name{fct_weighted_ll}
\alias{fct_weighted_ll}
\title{Compute weighted log likelihood}
\usage{
fct_weighted_ll(gamma)
}
\arguments{
\item{gamma}{posterior}
}
\value{
weighted log likelihood
}
\description{
Compute weighted log likelihood
}
