% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hthmix.R
\name{hthmix}
\alias{hthmix}
\title{Main mixrrr function}
\usage{
hthmix(
  x,
  y,
  k,
  lam = NULL,
  rank = NULL,
  chains = 50,
  maxiter = 100,
  verbose = TRUE,
  val_frac = 0.2,
  penal_search = 1:20/20
)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{number of clusters}

\item{lam}{penalization parameter}

\item{rank}{rank}

\item{chains}{number of chains}

\item{maxiter}{maximum number of iterations}

\item{verbose}{print progress?}

\item{val_frac}{validation fraction}

\item{penal_search}{penalization search parameter}
}
\value{
results
}
\description{
Main mixrrr function
}
