% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hthmix_cv.R
\name{hthmix_cv}
\alias{hthmix_cv}
\title{Cross Validation for Rank/K Selection in HTHMix}
\usage{
hthmix_cv(
  x,
  y,
  k_min = 2,
  k_max = 8,
  r_min = 1,
  r_max = 2,
  lam = NULL,
  chains = 50,
  maxiter = 100,
  verbose = TRUE,
  val_frac = 0.2,
  penal_search = 1:20/20
)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k_min}{minimum k to search}

\item{k_max}{maximum k to search}

\item{r_min}{minimum rank to search}

\item{r_max}{maximum rank to search}

\item{lam}{penalization parameter}

\item{chains}{number of chains}

\item{maxiter}{maximum iterations}

\item{verbose}{TRUE or FALSE}

\item{val_frac}{fraction for validation}

\item{penal_search}{penalty search vector}
}
\value{
list of results
}
\description{
Cross Validation for Rank/K Selection in HTHMix
}
