% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gamma.R
\name{fct_gamma}
\alias{fct_gamma}
\title{Compute Posterior}
\usage{
fct_gamma(
  x,
  y,
  k,
  N,
  clust_assign,
  selection,
  alpha,
  beta,
  y_sparse,
  rank,
  max_rank
)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{number of clusters}

\item{N}{doc}

\item{clust_assign}{doc}

\item{selection}{selection method}

\item{alpha}{sparsity tuning parameter}

\item{beta}{sparsity tuning parameter}

\item{y_sparse}{treat y as sparse?}

\item{rank}{force to be known rank}

\item{max_rank}{limit rank}
}
\value{
doc
}
\description{
Compute Posterior
}
