% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gamma.R
\name{fct_gamma}
\alias{fct_gamma}
\title{Compute Posterior}
\usage{
fct_gamma(x, y, k, N, p, m, lam, rank, clust_assign, val_frac, penal_search)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{clusters}

\item{N}{observations}

\item{p}{dimension}

\item{m}{dimension}

\item{lam}{penalty param}

\item{rank}{rank}

\item{clust_assign}{current assignment}

\item{val_frac}{fraction validation}

\item{penal_search}{search vector}
}
\value{
tibble of posterior
}
\description{
Compute Posterior
}
