% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hthmix_comp.R
\name{fct_hthmix_comp}
\alias{fct_hthmix_comp}
\title{Compute mixrrr}
\usage{
fct_hthmix_comp(
  x,
  y,
  k,
  maxiter,
  lam,
  rank,
  clust_assign,
  val_frac,
  penal_search
)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{clusters}

\item{maxiter}{maximum number of iterations}

\item{lam}{penalization parameter}

\item{rank}{rank}

\item{clust_assign}{initial cluster assignment}

\item{val_frac}{validation fraction}

\item{penal_search}{penalization search vector}
}
\value{
results
}
\description{
Compute mixrrr
}
