% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_hthmix_comp.R
\name{fct_hthmix_comp}
\alias{fct_hthmix_comp}
\title{Compute mixrrr}
\usage{
fct_hthmix_comp(x, y, k, maxiter, clust_assign, selection, alpha, beta)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{number of clusters}

\item{maxiter}{maximum number of iterations}

\item{clust_assign}{initial cluster assignment}

\item{selection}{selection method}

\item{alpha}{sparsity tuning parameter}

\item{beta}{sparsity tuning parameter}
}
\value{
results
}
\description{
Compute mixrrr
}
