% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate_run.R
\name{fct_simulate_run}
\alias{fct_simulate_run}
\title{Simulate and run mixrrr}
\usage{
fct_simulate_run(params)
}
\arguments{
\item{params}{list of parameters for simulation}
}
\value{
list of results
}
\description{
Simulate and run mixrrr
}
