% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hthmix.R
\name{hthmix}
\alias{hthmix}
\title{Main mixrrr function}
\usage{
hthmix(
  x,
  y,
  k,
  nstart = 1,
  init_assign = NULL,
  selection = c("universal", "cv"),
  alpha = 2 * sqrt(3),
  beta = 1,
  maxiter = 1000,
  verbose = TRUE
)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}

\item{k}{number of clusters}

\item{nstart}{number of times to initialize}

\item{init_assign}{initial assignment}

\item{selection}{selection method}

\item{alpha}{sparsity tuning parameter}

\item{beta}{sparsity tuning parameter}

\item{maxiter}{maximum number of iterations}

\item{verbose}{print progress?}
}
\value{
results
}
\description{
Main mixrrr function
}
